requirements.txt
flask
flask-cors
torch
transformers
Pillow